Transaction should not block opening Realm
This is the test case only, the problem has been fixed when we introduce
the schema caches.

Since we are caching the schemas for typed Realm, when a transaction
lock is holding by one type Realm, open Realm instance in another thread
should not be blocked.
But if there is no typed Realm instance opened, DynamicRealm transaction
should block the typed Realm openning.