Remove HandlerController from the public API. (#2632)
HandlerController was made public by mistake. This is an class only used by internal API's, so shouldn't be public.

Most likely it was public due to a testing class (HandlerProxy) in a separate package needing access to it. HandlerProxy was instead moved to io.realm to work around this as keeping the public API clean has priority.