Better exception message when using custom migrations (#2715)
Custom migration classes has been a source of quite a few support questions. This commit adds an extra check for that specific case in order to generate a better exception message that also points to a fix.

The alternative was to ignore the Migration class when comparing RealmConfigurations, and even though that would be very unlikely it could also lead to some very subtle bugs. By doing it this way, we stick to the normal semantics of Java equality.