Workaround to add android.jar for transform (#2741)
Close #2703

While javassit manipulating the method, it requires all of the classed
used in the method can be found in the class pool. But android.jar will
not be passed to the transform -- It is not in any of
QualifiedContent.Scope definitions.

We use project.android.bootClasspath which is public right now to get
the path to android.jar. Project is passed from gradle.
See https://android.googlesource.com/platform/tools/base/+/studio-master-dev/build-system/gradle/src/main/groovy/com/android/build/gradle/BaseExtension.java#737