Add RealmInvalidDatabaseException
This is necessary.
For some reasons, the db file might get corrupted (the most reasonable
thing is final user install an old version of app which tries to open a
Realm db created by newer version of Realm). The only way to solve that
is deleting the Realm file and start over.

Instead of asking user to catch IllegalStateException and checking the
message, throw a special exception for the purpose is more feasible.