Fix listeners can exposing unsynchronized RealmResults (Async queries) (#2951)
When async queries are updated, they call listeners in a specific order. Before this commit this had a very subtle bug, so accessing synchronous RealmResults from inside a async RealmResult listener might hit a detached row accessor.

Reason being is that the Realm is advanced, then all listeners called in order, but RealmResults are not synchronised until just before listener is called.

This commit fixes this so all RealmResults are now synced before the listener is called.