Race condition between Realm change notifications and UI events (#2990)
This commit fixes a bug surfaced by our Realm Android Adapter example. Before this commit we posted all REALM_CHANGED events to the looper queue as normal. This meant that potential touch input events could get in front which in turn could cause a relayout of a ListView which would then crash because it detected a difference between it's cached adapter size and the actual one.

This commit now forces all local commits to be placed at the front of the queue eliminating that risk.