Fix async transaction problem when async query already exist. (#2780)
I have rewritten how async transactions works. Before they did this:

1) commitOnBackgroundThread() -> post REALM_CHANGED -> post onSuccess runnable

The problem with that approach was that the REALM_CHANGED would be swallowed if async queries existed which meant that the async transaction would call onSuccess on an old version of the Realm (making it look like it didn't work).

Instead we now do this:

2) commitOnBackgroundThread -> post runnable that calls HandlerController.handleAsyncTransactionCompleted(runnable)

The special runnable is treated as a combined REALM_CHANGED + Callback, which makes it possible for us to queue up callbacks until we are finally able to trigger all of them.

Unfortunately the Handler has poor support for this so it means that we no longer can detect if such a message is in the queue. This means that it introduces a slight chance of such a event plus a real REALM_CHANGED event to be in the message queue at the same time. I considered that acceptable (since it can already happen today), and since preventing this will introduce more complexity to something that is already entirely to complex.