Fix notification tests for following changes (#3301)
* Fix notification tests for following changes

To support multiprocess & OS notifications, a dedicated thread will
monitor the Realm changes and send notifications to other thread. This
means our tests cannot assume notifications will be sent for every
transaction commit.

Also, the Realm.handler will moved to a lower layer, it's better to not
rely on it in the test as well.

To avoid a massive PR, tests changes are made separately first.