Wait all async tasks done before next test
This is highly related with #1900.
Below things are still guaranteed by this change after commit async
transaction:
* When callback function called (onSuccess/onError), the background
  Realm is closed.
* When any change listeners called, the background Realm is closed.

What is true now but it is not guaranteed in the future:
* Background Realm might not be closed before REALM_CHANGED sent (not
  received).

Due to this, to avoid the flaky tests, we have to ensure all async tasks
quit peacefully before start the next test. This is implemented by wait
and check in the TestRealmConfigurationFactory.

NOTE: Any test, if the async tasks cannot be finished peacefully in a
certain time, it has to be considered as a problem and fixed.

This would be needed by OS notifications since Java won't have a precise
control of sending REALM_CHANGED anymore.