Remove improper incompatibleLockFile test (#3324)
This test is not constructed properly. It tries to modify the lock file
while the Realm is still opened. This would cause some undefined
behaviour in core. Especially when we are using the generic
external_commit_helper which is based on SharedGroup::wait_for_change()
the helper thread will freeze because of polling hangs forever.

Just remove the test, core should ensure the proper exception thrown and
we have other tests to ensure the exception conversion right.