Integrate Object Store [PART4] - OS notifications (#3370)
* Use OS's notification mechanism to notify threads.
* Create RealmNotificer interface for decouple Android related handler
  logic.
* Create AndroidNotifier for the handler notifications.

The major change of this PR is about the timing. The notifications are
not sent immediately after transaction committed. Instead, there is a
dedicated thread monitoring changes and notify others when changes
happen.

The known problem is for every RealmConfiguration, a monitor thread will
be created which is not ideal for app which is using multiple
RealmConfiguration.

There are different implementations for the monitoring thread in OS. For
Android, we can choose from generic which is based on the core's
wait_for_change() and android which is used by dotnet based on the
named pipe.
To align with dotnet, we are using the named pipe for now which also
enables notifications between realm-java and realm-dotnet.