Check if mediator misses any class
This is found when #3343.
If a open a existing Realm file with a mediator doesn't have necessary
link object defined, it should throw. This is not a critical problem
since it will throw any way when user try to operate with that missing
class by the typed Realm interface, but throwing earlier is always a
good thing to do.