Single daemon thread for notification
With changes in https://github.com/realm/realm-object-store/pull/197 ,
there is only one thread will be create for listening changes for all
different Realms. But an additional SharedGroup will be created in the
daemon thread for determine which SharedGroup has changed since last
time. That requires some changes in the java side especially the daemon
thread should not be created when Realm.compactRealm called.