Fix the warnings of StrictMode
The StrictMode warnings is caused by testInMemoryRealm which try to
detect if there are IOs in the thread when using in-memory Realm.

Well, that test is created by me and it doesn't make any sense to check
with StrictMode.
In memory Realm doesn't mean there will be no disk IOs, it only means
if all the instances are closed, the data will not be persisted on the
disk.

And the StrictMode warning is cause by the last statement
`StrictMode.enableDefault()`. StrictMode is not turned on for testing by
default!

Fix #3141