Get findAllAsync and findFirstAsync back
the findAllAsync got deprecated since from OS Results point of view,
the query updating will run in the background by default, then there
would be no difference between findAll and findAsync. Apparently the
isLoad() and load() should be deprecated as well. So like cocoa, the
Query will be executed immediately if user tries to access the
element in the Results.

But ... That was wrong because of those use cases:

1) RxJava support, in the subscription, it needs to check isLoaded()
   to determine the next step. And it will be fired once after the
   Observable created. By always returning true from isLoaded(),
   that means for RxJava will always run sync query at the first time.
2) Create a RealmResults and pass it to a list adapter. Since UI will
   always call size() which will run the query immediately and the
   first time query becomes synced.

So, we get all async related APIs back and keep them having the same
functionality like before. The behavior won't be exact the same, but
from user's perspective, they are the same as before.