Add OrderedRealmCollectionSnapshot (#4172)
Introduce OrderedRealmCollectionSnapshot as another type of
OrderedRealmCollection.
A snapshot can be created from RealmResults or RealmList.
A snapshot is backed by a snapshot of OS Results. So the snapshot
itself won't be updated. The size and order stay the same forever
(elements inside are still live objects.).

Since the RealmResults is auto-updated all the time, snapshot will be
usefull when changing the results in a simple loops.

This commit also moves the common code from RealmResults to
OrderedRealmCollectionImpl since those can be shared with snapshot
implementation.

Implement #3883