findFirstAsync returns invalid row if no object
Fix #4352
The findFirstAsync()'s behavior doesn't match the javadoc from the first
day the API was introduced. It kept running the query until it could
find a row match the query condition. This behavior create difficulties
if user want to check if there is no object in the db. Also it was not
consistent with the behavior of findFirst() which will return an invalid
row in the same condition.