Work around the memmove bug on Samsung device (#4402)
There was a bug for memmove on some Samsung devices. The functions
returns "dest + n" instead of "dest".
This has been identified earlier by QT. See:
https://bugreports.qt.io/browse/QTBUG-34984
The relevant android bug:
https://code.google.com/p/android/issues/detail?id=81692

This fix try to test if the device have this issue first, if yes, switch
to a own implementation of memmove.

This fix works for most cases, but since the memmove bug existing in the
system, there would be other lib/system call the buggy version of
memmove which could still corrupt the memory -- which means there are
still some possibilities that app gets crashed on those devices.