Implement getInstanceAsync (#4570)
Fix #2299
- Add APIs to get Realm instance asynchronously.
- Remove some useless methods.

There some work need to be done before create the first Realm instance
in the process, like creating schema table, doing migration, etc.. Those
could block the UI thread quite badly. This commit tries to do those
initialization work in the background and hold a Realm instance in the
background until the 2nd instance created in the caller thread.

A better solution than this would be do initialization in the
background and only deliver a column indices cache to caller thread
without holding a Realm instance in the background. But that is not
possible since from the current database design, we cannot know if the
schema changes compared with the last time it was opened.

Also create a SharedGroup in the background and handover it to the
caller thread is not ideal as well. That not only requires some design
changes in the Object Store RealmCoordinator, but also is a very
special use case of SharedGroup which core is not designed for.
SharedGroup Leaking during the handover is another flaw for this
solution -- we can only rely on the GC to collect the leaked SharedGroup
during handover then.