Let Object Store handle table creations
This tries to progressively move more things about schemas to Object
Store.

First the concept of Schema in Object Store is not the same as what we
have in Java. It is very much just a schema information holder and won't
take care of the schema modifications. That says it is more like the
ColumnIndices cache in the Java binding.
So this commit try to:
- Instead of inheriting from the RealmSchema, change the
OsRealmSchema/OsRealmObjectSchema to OsSchemaInfo/OsObjectSchemaInfo.
They behave as a simple Java wrapper to the relevant OS objects.
- Add functions to Proxy classes which will create its own
OsObjectSchemaInfo and those info can be used to create a OsSchemaInfo
through the mediator.
- Call `SharedRealm.updateSchema` with the OsSchemaInfo which is got
from the proxy interface to do table initialization.
 - This will also fix a minor bug we have before, All tables are created
 even if the class is not in the module.
- Migration is still handled in the old way, and it will be solved in
the future, to let Object Store handle it.
- ColumnIndices are still kept for now, but it should be replaced by the
OsSchemaInfo/OsObjectSchemaInfo in the future.