Clean up pre-null related tests created in 0.83.0
From 0.83.0, we begin to support the nullable fields. But because of a
bug, the Realm file which was created in 0.82.2 contains all schemas
defined in the project instead of those only defined in the RealmModule.
This will cause lots of troubles after we fix this bug and move to
Object Store's schema.

- Create string-only-pre-null-0.82.2.realm with 0.82.2 which only
  contains StringOnly class.
- Create string-only-required-pre-null-0.82.2.realm with 0.82.2 which
  only contains StringOnlyRequired class.
- Use dynamic API to rewrite migrationException_realmListChanged.
- Remove default-before-migration.realm.