Link to debug core lib for debug build (#4221)
Ideally we should run CI testing with debug enabled for core. But
because of the assertions slow down the testing too much (1 hour
for one flavour testing), we just add a opt-out gradle option to enable
it when needed. So CI would just have the same speed as before by
linking with release core.

But I suggest we enable it on our develop env all the time. Since a lot
of issues will just assert earlier and all the debug symbols in core make
debugging much easier for daily life.

just add below to your ~/.gradle/gradle.properties
enableDebugCore=true