Primitive List feature(Part5: Element type check in annotation processor) (#5168)
* @NotNull as alias for @Required (#5161)

This PR adds support for org.jetbrains.annotations.NotNull as an alias for @Required. This means that Realm now understands the Kotlin type-system in the Realm schema.

Previously we reported an error if you did @Required RealmList<MyType> list = new RealmList<>();. It is a bit unclear why, since a RealmList is always non-null on managed objects. Since it made the interop with Kotlin kinda strange, I made a change so you can now do @Required RealmList<MyType> list = new RealmLis<>() as well as val list : RealmList<MyType> = RealmList();

After this change it is no longer possible to do val person : Person either, you have to do val person : Person?. This reflects the constraints of Realm, but if people are able to maintain the variant themselves (hint: not possible when using sync), then they can use a custom non-null getter.

* implement elemenrt type check of RealmList

* fix error message

* remove extra line

* fix typo

* address review comments