Refactor JNI Java arrays accessors
The original implementation has a serious memory issue that the
ReleaseXxxArrayElements might be called before accessing if the
copy/move constructor is invoked. That could happen when use
JniArraysOfArray where the JniXxxArray would be copied when push it to
the vector. The JNI SortDescriptor may trigger various native crashes
because of using this.