Add realm.ignoreKotlinNullability
To disable treating kotlin non-null types as @Required.
From v3.6.0, all the kotlin RealmModel's non-null field will be set as
required in the schema. However it is a breaking change for the kotlin
project which has a Realm file created before 3.6.0.

To disable this behaviour introduced in 3.6.0, add below things to the
project's build.gradle:

kapt {
  arguments {
    arg("realm.ignoreKotlinNullability", true)
  }
}

Close #5412