Fix a socket timeout issue with test server
If startRealmObjectServer was called twice without stop in between,
syncServerChildProcess will be overwritten by a terminated process since
the 2nd call of creating sync server will fail and set the syncServerChildProcess.
This would make the stopRealmObjectServer get stuck forever.

In practice, we need to make sure the server has been stopped before
start it. So just treat the "start" as a "restart" command which will
stop the sync server if needed.

Thus, no need to stop in tear down functions in java side anymore.